requirements.txt
pymupdf
python-docx
google-generativeai